Delete unnecessary parts under .github
This is a fork to satisfy security requirements, we're never going to do
any work in this repo. If we ever need to make any changes we'll
upstream them to the original repository.

The build workflow is useful to keep, since it confirms that the dist js
file hasn't been tampered with.